+++ form.css
@@ -3,7 +3,6 @@ body {
     font-weight: bold;
 }
 
-
 main {
     height: 900px;
     width: 500px;
