+++ index.php
@@ -1,7 +1,5 @@
 
 <?php require 'pull_data.php';?>
-
-
 <html lang="en">
 <head>
     <meta charset="UTF-8">
