+++ push_data.php
@@ -1,5 +1,4 @@
 <?php
-
 require 'functions.php';
 
 $db = new PDO('mysql:host=127.0.0.1;dbname=portfolio', 'root');
